// components/click-up/src/actions/time-tracking/types/TimeEntriesDateRangeQueryParams.ts
export interface TimeEntriesDateRangeQueryParams {
  start_date?: string;
  end_date?: string;
  assignee?: string;
  include_task_tags?: string;
  include_location_names?: string;
  space_id?: string;
  folder_id?: string;
  list_id?: string;
  task_id?: string;
  custom_task_ids?: string;
  team_id?: string;
}